requirements.txt
uvloop
aiohttp>=1.0
coloredlogs
requests
github3.py
pygit2
pyyaml
pytest
